+++ software/firmware/oneway_common.h
@@ -79,7 +79,7 @@ struct pp_tag_poll  {
 	uint8_t message_type;                   // Packet type identifier so the anchor knows what it is receiving.
 	uint8_t subsequence;                    // Index of which broadcast sequence number this packet is.
 	uint8_t reply_after_subsequence;        // Tells anchor which broadcast subsequence number to respond after.
-	uint16_t anchor_reply_window_in_us;     // How long each anchor response window is. Each window allows multiple anchor responses.
+	uint32_t anchor_reply_window_in_us;     // How long each anchor response window is. Each window allows multiple anchor responses.
 	uint16_t anchor_reply_slot_time_in_us;  // How long that slots that break up each window are.
 	uint16_t anchor_reply_num_slots;
 	struct ieee154_footer footer;
